scripts/requirements.txt
# Requirements for homelab management scripts

# Docker Python SDK for container management
docker>=6.1.0

# HTTP requests for API calls
requests>=2.31.0

# YAML parsing for compose files
PyYAML>=6.0.1

# Type checking and annotations
typing-extensions>=4.7.0

# Optional: Enhanced terminal output
rich>=13.4.0

# Optional: Progress bars for long operations
tqdm>=4.65.0

# Optional: Better date/time handling
python-dateutil>=2.8.2
